+++ CHANGELOG.md
@@ -4,8 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+### Added
+- Use `Honeybadger.metricsHandler` to send us request metrics!
 
 ## [1.0.0] - 2016-04-21
-
 ### Changed
 - Entirely new API. See the [README](README.md) for details!